remove mobile app id
