+++ plugins/MobileAppMeasurable/config/test.php
@@ -1,0 +1,7 @@
+<?php
+
+return array(
+
+    'Piwik\Plugins\MobileAppMeasurable\Type' => DI\object('Piwik\Plugins\MobileAppMeasurable\tests\Framework\Mock\Type'),
+
+);
