+++ tests/PHPUnit/Integration/Measurable/MeasurableTest.php
@@ -9,6 +9,7 @@
 namespace Piwik\Tests\Integration\Measurable;
 
 use Piwik\Db;
+use Piwik\Plugins\MobileAppMeasurable\tests\Framework\Mock\Type;
 use Piwik\Plugins\MobileAppMeasurable\Type as MobileAppType;
 use Piwik\Plugin;
 use Piwik\Measurable\Measurable;
@@ -84,7 +85,8 @@ class MeasurableTest extends IntegrationTestCase
     public function provideContainerConfig()
     {
         return array(
-            'Piwik\Access' => new FakeAccess()
+            'Piwik\Access' => new FakeAccess(),
+            'Piwik\Plugins\MobileAppMeasurable\Type' => new Type()
         );
     }
